Implemented CommandParserTest
